# src/main/resources/swagger.yaml
openapi: 3.0.0
servers:
  - description: SurfsUp Swagger
    url: /v1
info:
  description: |
    End-points for SurfsUp.
  version: "v1"
  title: Swagger SurfsUp
  contact:
    email: amanmanocha@yahoo.com
...
paths:
...
  /users/changePassword:
    post:
      tags:
        - users
      summary: Changes user's password
      operationId: changePassword
      responses:
        '200':
          description: successful operation
      '400':
        description: Authentication failed for user
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
          required: true
  /users/resetPassword:
    post:
      tags:
        - users
      summary: Resets user's password
      operationId: changePassword
      responses:
        '200':
          description: successful operation
      '400':
        description: Authentication failed for user
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
          required: true
...
  /users/deviceToken:
    put:
      tags:
        - users
      summary: Updates users device token
      operationId: updateDeviceToken
      responses:
        '200':
          description: successful operation
        '404':
          description: User not found
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserDeviceTokenRequest'
          required: true
      security:
        - bearerAuth: [ ]
  /users/sendOTP:
    post:
      tags:
        - users
      summary: Sends OTP to user
      operationId: updateDeviceToken
      responses:
        '200':
          description: successful operation
        '404':
          description: User not found
      security:
        - bearerAuth: [ ]
  /users/verifyEmail:
    post:
      tags:
        - users
      summary: Verifies email
      operationId: verifyEmail
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '404':
          description: Invalid or expired OTP
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OTP'
          required: true
      security:
        - bearerAuth: [ ]
  /users/logout:
    post:
      tags:
        - users
      summary: Logs out user
      operationId: logoutUser
      responses:
        '200':
          description: successful operation
        '404':
          description: User not found
      security:
        - bearerAuth: [ ]
...
components:
  schemas:
...
    ChangePasswordRequest:
      type: object
      properties:
        email:
          type: string
        oldPassword:
          type: string
        newPassword:
          type: string
        deviceType:
          type: string
    ResetPasswordRequest:
      type: object
      properties:
        email:
          type: string
        deviceType:
          type: string
...
    UpdateUserDeviceTokenRequest:
      type: object
      properties:
        deviceToken:
          type: string
    OTP:
      type: object
      properties:
        otp:
          type: string
...
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        deviceToken:
          type: string
        deviceType:
          type: string
        startTrialAt:
          type: integer
        endTrialAt:
          type: integer
        userType:
          type: string
        snoozeTill:
          type: integer
        disableAllAlerts:
          type: boolean
        notificationsPerHour:
          type: integer
        lastPaymentAt:
          type: integer
        nextPaymentAt:
          type: integer
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT